Update postgres types to match the language types: `int64<>BIGINT`, `uint32<>INTEGER`

+++ internal/db/migrations/2025-06-10.2-create_indexer_table_transactions.sql
@@ -3,7 +3,7 @@
 -- Table: transactions
 CREATE TABLE transactions (
     hash TEXT PRIMARY KEY,
-    to_id INTEGER NOT NULL,
+    to_id BIGINT NOT NULL,
     envelope_xdr TEXT,
     result_xdr TEXT,
     meta_xdr TEXT,
