+++ internal/db/migrations/2025-06-10.4-create_indexer_table_state_changes.sql
@@ -9,7 +9,7 @@ CREATE TABLE state_changes (
     ledger_created_at TIMESTAMPTZ NOT NULL,
     ledger_number INTEGER NOT NULL,
     account_id TEXT NOT NULL REFERENCES accounts(stellar_address),
-    operation_id INTEGER NOT NULL REFERENCES operations(id),
+    operation_id BIGINT NOT NULL REFERENCES operations(id),
     tx_hash TEXT NOT NULL REFERENCES transactions(hash),
     token TEXT,
     amount TEXT,
@@ -19,7 +19,7 @@ CREATE TABLE state_changes (
     key_value JSONB,
     offer_id TEXT,
     signer_account_id TEXT,
-    signer_weight INTEGER,
+    signer_weight BIGINT,
     spender_account_id TEXT,
     target_account_id TEXT,
     thresholds JSONB
